+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.types.js
@@ -1,6 +1,7 @@
 // @flow
 import type { ProgramStage } from '../../../metaData';
 import type { WidgetEffects, HideWidgets } from '../common/EnrollmentOverviewDomain';
+import type { UserFormField } from '../../FormFields/UserField';
 import type { LinkedRecordClick } from '../../WidgetsRelationship/WidgetTrackedEntityRelationship';
 import type {
     PageLayoutConfig,
@@ -33,6 +34,14 @@ export type PlainProps = {|
     onHandleScheduleSave: (eventData: Object) => void,
     pageStatus: string,
     eventStatus?: string,
+    eventAccess: {|
+        read: boolean,
+        write: boolean,
+    |} | null,
+    getAssignedUserSaveContext: () => { event: ApiEnrollmentEvent },
+    assignee: UserFormField | null,
+    onSaveAssignee: (newAssignee: UserFormField) => void,
+    onSaveAssigneeError: (prevAssignee: UserFormField | null) => void,
 |};
 
 export type Props = {|
@@ -41,6 +50,7 @@ export type Props = {|
     teiId: string,
     enrollmentId: string,
     orgUnitId: string,
-    eventId: string,
+    event: ApiEnrollmentEvent,
+    enrollmentSite: ApiEnrollment,
     initMode?: string,
 |};