+++ src/core_modules/capture-core/components/Pages/ViewEvent/RightColumn/AssigneeSection/index.js
@@ -1,4 +1,2 @@
 // @flow
-export { actionTypes as assigneeSectionActionTypes } from './assigneeSection.actions';
 export { AssigneeSection } from './AssigneeSection.container';
-export { saveAssigneeEpic } from './saveAssignee.epic';